Introduce base NotifyMailer to to abstract away shared Notify mail functionality
